refactor(test): fold #7449 math/code coverage into the sibling suite
Review feedback on this PR flagged review-checklist §14 R7 (test
budget): a suite for processMarkdownContent already exists at
MarkdownRendererSecurity.vitest.mts, so standing up a second file for
the same module duplicated its renderToDocument helper verbatim.
Fold the three new cases into that file as a second describe block
and trim the doc-comment rationale to a one-line pointer at the PR
description, per CLAUDE.md's guidance to keep process narrative out
of code comments.
